+++ src/MapEditor/MapEditor.cpp
@@ -1478,8 +1478,8 @@ void MapEditor::drawMap(ScreenBorder* pScreenborder, bool bCompleteMap) const {
             //draw map[x][y]
             SDL_Rect source = { (tile % NUM_TERRAIN_TILES_X)*zoomedTilesize, (tile / NUM_TERRAIN_TILES_X)*zoomedTilesize,
                                 zoomedTilesize, zoomedTilesize };
-            SDL_FRect drawLocation = {   pScreenborder->world2screenX(x*TILESIZE), pScreenborder->world2screenY(y*TILESIZE),
-                                        zoomedTilesize, zoomedTilesize };
+            SDL_FRect drawLocation = {   static_cast<float>(pScreenborder->world2screenX(x*TILESIZE)), static_cast<float>(pScreenborder->world2screenY(y*TILESIZE)),
+                                        static_cast<float>(zoomedTilesize), static_cast<float>(zoomedTilesize) };
             Dune_RenderCopyF(renderer, terrainSprite, &source, &drawLocation);
         }
     }
@@ -1498,7 +1498,9 @@ void MapEditor::drawMap(ScreenBorder* pScreenborder, bool bCompleteMap) const {
 
             SDL_Rect source = {static_cast<int>(Tile::TERRAINTILETYPE::TerrainTile_Slab) * zoomedTilesize, 0, zoomedTilesize,
                                zoomedTilesize};
-            SDL_FRect dest = { pScreenborder->world2screenX(position.x*TILESIZE), pScreenborder->world2screenY(position.y*TILESIZE), zoomedTilesize, zoomedTilesize };
+            SDL_FRect dest   = {static_cast<float>(pScreenborder->world2screenX(position.x * TILESIZE)),
+                              static_cast<float>(pScreenborder->world2screenY(position.y * TILESIZE)),
+                              static_cast<float>(zoomedTilesize), static_cast<float>(zoomedTilesize)};
 
             Dune_RenderCopyF(renderer, terrainSprite, &source, &dest);
 
@@ -1511,7 +1513,9 @@ void MapEditor::drawMap(ScreenBorder* pScreenborder, bool bCompleteMap) const {
                 for(int x = position.x; x < position.x+2; x++) {
                     SDL_Rect source = {static_cast<int>(Tile::TERRAINTILETYPE::TerrainTile_Slab) * zoomedTilesize, 0, zoomedTilesize,
                                        zoomedTilesize};
-                    SDL_FRect dest = { pScreenborder->world2screenX(x*TILESIZE), pScreenborder->world2screenY(y*TILESIZE), zoomedTilesize, zoomedTilesize };
+                    SDL_FRect dest   = {static_cast<float>(pScreenborder->world2screenX(x * TILESIZE)),
+                                      static_cast<float>(pScreenborder->world2screenY(y * TILESIZE)),
+                                      static_cast<float>(zoomedTilesize), static_cast<float>(zoomedTilesize)};
 
                     Dune_RenderCopyF(renderer, terrainSprite, &source, &dest);
                 }
@@ -1574,7 +1578,7 @@ void MapEditor::drawMap(ScreenBorder* pScreenborder, bool bCompleteMap) const {
             const auto* const WallSprite = pGFXManager->getZoomedObjPic(ObjPic_Wall, currentZoomlevel);
 
             SDL_Rect source = { maketile * zoomedTilesize, 0, zoomedTilesize, zoomedTilesize };
-            SDL_FRect dest = { pScreenborder->world2screenX(position.x*TILESIZE), pScreenborder->world2screenY(position.y*TILESIZE), zoomedTilesize, zoomedTilesize };
+            SDL_FRect dest = { static_cast<float>(pScreenborder->world2screenX(position.x*TILESIZE)), static_cast<float>(pScreenborder->world2screenY(position.y*TILESIZE)), static_cast<float>(zoomedTilesize), static_cast<float>(zoomedTilesize) };
 
             Dune_RenderCopyF(renderer, WallSprite, &source, &dest);
 
@@ -1611,7 +1615,9 @@ void MapEditor::drawMap(ScreenBorder* pScreenborder, bool bCompleteMap) const {
             Coord frameSize = world2zoomedWorld(getStructureSize(structure.itemID)*TILESIZE);
 
             SDL_Rect source = { frameSize.x*(structure.itemID == Structure_WindTrap ? 9 : 2), 0, frameSize.x, frameSize.y };
-            SDL_FRect dest = { pScreenborder->world2screenX(position.x*TILESIZE), pScreenborder->world2screenY(position.y*TILESIZE), frameSize.x, frameSize.y };
+            SDL_FRect dest   = {static_cast<float>(pScreenborder->world2screenX(position.x * TILESIZE)),
+                              static_cast<float>(pScreenborder->world2screenY(position.y * TILESIZE)),
+                              static_cast<float>(frameSize.x), static_cast<float>(frameSize.y)};
 
             Dune_RenderCopyF(renderer, ObjectSprite, &source, &dest);
